feat: add panic recovery to stream handling and improve error management

+++ rpcruntime/errors_test.go
@@ -8,7 +8,7 @@ import (
 
 func resetForTest() {
 	registryMu.Lock()
-	registry = make(map[int]errorRecord)
+	registry = make(map[int64]errorRecord)
 	registryMu.Unlock()
 	nextErrorID.Store(0)
 	startCleanerOnce = sync.Once{}
@@ -70,7 +70,7 @@ func TestConcurrencySafety(t *testing.T) {
 	storeCount := 200
 	wg.Add(storeCount)
 
-	ids := make([]int, storeCount)
+	ids := make([]int64, storeCount)
 	for i := 0; i < storeCount; i++ {
 		i := i
 		go func() {
